Implement mark as done function and task class
Because,
* the old method requires a more complicated 2-dimensional array.
* mark as done is a common behavior for all tasks (isDone).
* there is flexibility for future additions of other functions and
parameters associated with the task (object).

The mark as done done feature is added to allow users to mark their
tasks as done. Instead of using a 2-dimensional array which is
complicated and inflexible, I changed the storage of tasks from array
(String) to class (Task). Task contains a description of the task
(description) and a boolean to track whether a task is completed
(isDone). Primarily, this commit implements the mark as done command
and changed the tasks from string to an object.

This commit should make future implementation of other functions easier.