Add Increment A-CodeQuality
Refactor by moving methods from the main Duke class to 2 other new classes, DukeController and UI.
Methods that handle printing to the console are moved to the UI class,while methods that handle the logic based on user input is moved to the DukeController class.
Additionally introduced constants in DukeController, UI and Task classes for increased clarity.