Add output task list to file
Because:
* the task list is only useful if it is persistent between uses.
* it allows the user to close the program between uses to save
resources.
* there is a reference if even if the program is not launched.

The format of the stored tasks are comma seperated. This is to allow easy
seperation of different parameters of each task, compared to the
regular display string. This would simplify the process of parsing the
file to the task list.

There are also small bug fixes to the program, making it more robust to
input syntax errors.